Split `serializers` into package grouped by use case
